+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flashing support for the STM32G0 series.
 - Flashing support for the STM32F0 series.
 - Added SWD support for the JLink.
+- Support for RISCV debugging using a Jlink debug probe.
 
 ### Changed